# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TipEase API - Cashless tipping platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      security:
        - clerkAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
    put:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/me/role:
    post:
      operationId: setUserRole
      tags: [users]
      summary: Set user role
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetUserRoleBody"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # Wallet
...
  /events/{eventId}/tips:
    get:
      operationId: getEventTips
      tags: [tips]
      summary: Get tips for an event (vendor or host view)
      security:
        - clerkAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of tips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tip"

  # QR Code Scanning
...
  /tips:
    post:
      operationId: sendTip
      tags: [tips]
      summary: Send a tip
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendTipBody"
      responses:
        "201":
          description: Tip sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tip"
        "400":
          description: Insufficient balance or limit exceeded
  /tips/history:
    get:
      operationId: getTipHistory
      tags: [tips]
      summary: Get tip history for current user
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of tips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tip"

  # Vendors
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List approved vendors (searchable directory)
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendorProfile
      tags: [vendors]
      summary: Create or update vendor profile
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendorBody"
      responses:
        "201":
          description: Vendor profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/me:
    get:
      operationId: getMyVendorProfile
      tags: [vendors]
      summary: Get current user's vendor profile
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Vendor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/me/events:
    get:
      operationId: getVendorEvents
      tags: [vendors]
      summary: Get events assigned to this vendor
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"

  # Notifications
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [guest, host, vendor, admin]
        limitTier:
          type: string
          enum: [standard, verified, enhanced, custom]
        walletBalance:
          type: number
        stripeCustomerId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - email
        - name
        - role
        - limitTier
        - walletBalance
        - createdAt
    UpdateUserProfileBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
    SetUserRoleBody:
      type: object
      properties:
        role:
          type: string
          enum: [guest, host, vendor, admin]
      required:
        - role
...
    Event:
      type: object
      properties:
        id:
          type: string
        hostId:
          type: string
        vendorId:
          type: ["string", "null"]
        name:
          type: string
        eventType:
          type: string
          enum: [private_party, wedding, corporate, restaurant, public_event, other]
        venueName:
          type: string
        venueAddress:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        expectedGuests:
          type: ["integer", "null"]
        perGuestCap:
          type: ["number", "null"]
        status:
          type: string
          enum: [draft, pending_vendor, confirmed, active, ended, cancelled]
        vendorConfirmed:
          type: boolean
        hostName:
          type: string
        vendorName:
          type: ["string", "null"]
        totalTips:
          type: number
        tipCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - hostId
        - name
        - eventType
        - venueName
        - venueAddress
        - startTime
        - endTime
        - status
        - vendorConfirmed
        - hostName
        - totalTips
        - tipCount
        - createdAt
...
    SendTipBody:
      type: object
      properties:
        eventId:
          type: string
        qrToken:
          type: string
        amount:
          type: number
      required:
        - eventId
        - qrToken
        - amount
    Tip:
      type: object
      properties:
        id:
          type: string
        eventId:
          type: string
        guestId:
          type: string
        vendorId:
          type: string
        amount:
          type: number
        eventName:
          type: string
        vendorName:
          type: string
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - eventId
        - guestId
        - vendorId
        - amount
        - eventName
        - vendorName
        - status
        - createdAt
    Vendor:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        businessRegNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        rating:
          type: ["number", "null"]
        totalEvents:
          type: integer
        totalEarnings:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - companyName
        - contactName
        - email
        - status
        - totalEvents
        - totalEarnings
        - createdAt
    CreateVendorBody:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        phone:
          type: string
        businessRegNumber:
          type: string
      required:
        - companyName
        - contactName
...